+++ Readio/src/main/java/ninegle/Readio/book/controller/BookController.java
@@ -1,26 +1,15 @@
 package ninegle.Readio.book.controller;
 
-import java.util.List;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
-import ninegle.Readio.book.domain.BookSearch;
 import ninegle.Readio.book.dto.BookListResponseDto;
 import ninegle.Readio.book.dto.BookResponseDto;
-import ninegle.Readio.book.dto.BookSearchResponseDto;
-import ninegle.Readio.book.dto.ReviewListResponseDto;
-import ninegle.Readio.book.dto.ReviewRequestDto;
 import ninegle.Readio.book.service.BookService;
 import ninegle.Readio.global.unit.BaseResponse;
 
@@ -59,29 +48,6 @@ public class BookController {
 		return bookService.searchBooks(keyword, page, size);
 	}
 
-	@PostMapping("/{book_id}/reviews")
-	public ResponseEntity<BaseResponse<Void>> save(@RequestBody ReviewRequestDto review,
-		@PathVariable("book_id") Long bookId) {
-		return bookService.save(review, bookId);
-	}
-
-	@DeleteMapping("/{book_id}/reviews/{review_id}")
-	public ResponseEntity<BaseResponse<Void>> delete(@PathVariable("review_id") Long reviewId) {
-		return bookService.delete(reviewId);
-	}
-
-	@PutMapping("/{book_id}/reviews/{review_id}")
-	public ResponseEntity<BaseResponse<Void>> update(@RequestBody ReviewRequestDto review,
-		@PathVariable("review_id") Long reviewId) {
-		return bookService.update(review, reviewId);
-	}
-
-	@GetMapping("/{book_id}/reviews")
-	public ResponseEntity<BaseResponse<ReviewListResponseDto>> getReviews(@PathVariable("book_id") Long bookId
-		, @RequestParam(defaultValue = "1") int page,
-		@RequestParam(defaultValue = "3") int size) {
-		return bookService.getReviewList(bookId, page, size);
-	}
 
 
 	//파일 받아오기 예시 코드
