+++ Readio/src/main/java/ninegle/Readio/book/controller/PreferenceController.java
@@ -5,7 +5,6 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -13,10 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 import ninegle.Readio.book.dto.BookIdRequestDto;
-import ninegle.Readio.book.dto.BookPreferenceListDto;
-import ninegle.Readio.book.dto.ReviewListResponseDto;
-import ninegle.Readio.book.dto.ReviewRequestDto;
-import ninegle.Readio.book.service.BookService;
+import ninegle.Readio.book.dto.preferencedto.PreferenceListResponseDto;
 import ninegle.Readio.book.service.PreferenceService;
 import ninegle.Readio.global.unit.BaseResponse;
 
@@ -46,7 +42,7 @@ public class PreferenceController {
 	}
 
 	@GetMapping("/")
-	public ResponseEntity<BaseResponse<BookPreferenceListDto>> getPreferences(
+	public ResponseEntity<BaseResponse<PreferenceListResponseDto>> getPreferences(
 		@RequestParam(defaultValue = "1") int page,
 		@RequestParam(defaultValue = "3") int size) {
 		return service.getPreferenceList(page,size);
