+++ Readio/src/main/java/ninegle/Readio/book/dto/BookIdRequestDto.java
@@ -13,6 +13,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class BookIdRequestDto {
+
 	private Long id;
 
 }
