+++ Readio/src/main/java/ninegle/Readio/book/dto/reviewdto/ReviewListResponseDto.java
@@ -1,10 +1,11 @@
-package ninegle.Readio.book.dto;
+package ninegle.Readio.book.dto.reviewdto;
 
 import java.util.List;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import ninegle.Readio.book.dto.PaginationDto;
 
 /**
  * Readio - ReviewListResponseDto
