+++ Readio/src/main/java/ninegle/Readio/book/mapper/ReviewMapper.java
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Component;
 import ninegle.Readio.book.domain.Book;
 import ninegle.Readio.book.domain.Review;
 import ninegle.Readio.book.dto.PaginationDto;
-import ninegle.Readio.book.dto.ReviewListResponseDto;
-import ninegle.Readio.book.dto.ReviewRequestDto;
-import ninegle.Readio.book.dto.ReviewResponseDto;
-import ninegle.Readio.book.dto.ReviewSummaryDto;
+import ninegle.Readio.book.dto.reviewdto.ReviewListResponseDto;
+import ninegle.Readio.book.dto.reviewdto.ReviewRequestDto;
+import ninegle.Readio.book.dto.reviewdto.ReviewResponseDto;
+import ninegle.Readio.book.dto.reviewdto.ReviewSummaryDto;
 import ninegle.Readio.user.domain.User;
 
 /**
