+++ Readio/src/main/java/ninegle/Readio/book/service/BookService.java
@@ -1,6 +1,5 @@
 package ninegle.Readio.book.service;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -16,7 +15,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -42,14 +40,8 @@ import ninegle.Readio.global.exception.BusinessException;
 import ninegle.Readio.global.exception.domain.ErrorCode;
 import ninegle.Readio.user.service.UserContextService;
 import ninegle.Readio.user.service.UserService;
-import ninegle.Readio.user.domain.User;
 import ninegle.Readio.book.mapper.ReviewMapper;
-import ninegle.Readio.book.domain.Review;
 import ninegle.Readio.book.dto.PaginationDto;
-import ninegle.Readio.book.dto.ReviewListResponseDto;
-import ninegle.Readio.book.dto.ReviewRequestDto;
-import ninegle.Readio.book.dto.ReviewResponseDto;
-import ninegle.Readio.book.dto.ReviewSummaryDto;
 import ninegle.Readio.book.repository.ReviewRepository;
 import ninegle.Readio.global.unit.BaseResponse;
 
@@ -100,16 +92,6 @@ public class BookService {
 		return new ArrayList<>(result);
   }
 
-  private void updateRatingInBookSearch(long bookId) {
-		BigDecimal rating = reviewRepository.findAverageRatingByBook(bookId);
-		if (rating != null) {
-			rating = BigDecimal.ZERO;
-		}
-		BookSearch bookSearch = bookSearchRepository.findById(bookId)
-			.orElseThrow(() -> new BusinessException(ErrorCode.BOOK_NOT_FOUND));
-		bookSearch.updateRating(rating);
-		bookSearchRepository.save(bookSearch);
-  }
 
 	public ResponseEntity<BaseResponse<Void>> save(BookRequestDto request) {
 
@@ -212,54 +194,6 @@ public class BookService {
 	public Book getBookById(long id) {
 		return bookRepository.findById(id).orElseThrow(
 			() -> new NoSuchElementException());
-	}
-
-	public Review getReviewById(long id) {
-		return reviewRepository.findById(id).orElseThrow(
-			() -> new NoSuchElementException());
-	}
-
-	// Review Create
-	public ResponseEntity<BaseResponse<Void>> save(ReviewRequestDto reviewRequestDto, long book_id) {
-		User user = userService.getById(userContextService.getCurrentUserId());
-		Book book = getBookById(book_id);
-		updateRatingInBookSearch(book_id);
-		reviewRepository.save(reviewMapper.toEntity(reviewRequestDto, user, book));
-		return BaseResponse.ok("후기 등록이 정상적으로 수행되었습니다.",null, HttpStatus.CREATED);
-	}
-
-	// Review Delete
-	public ResponseEntity<BaseResponse<Void>> delete(Long reviewId) {
-		Review review = getReviewById(reviewId);
-		reviewRepository.delete(review);
-		updateRatingInBookSearch(review.getBook().getId());
-		return BaseResponse.ok("삭제가 성공적으로 수행되었습니다.", null,HttpStatus.OK);
-	}
-
-	// Review Update
-	public ResponseEntity<BaseResponse<Void>> update(ReviewRequestDto reviewRequestDto, Long reviewId) {
-		Review review = getReviewById(reviewId);
-		reviewRepository.save(reviewMapper.updateEntity(review, reviewRequestDto));
-		updateRatingInBookSearch(review.getBook().getId());
-		return BaseResponse.ok("후기 수정이 정상적으로 수행되었습니다.", null,HttpStatus.OK);
-	}
-
-	public ResponseEntity<BaseResponse<ReviewListResponseDto>> getReviewList(Long bookId, int page, int size) {
-
-		Book book = getBookById(bookId);
-		Pageable pageable = PageRequest.of(page - 1, size);
-		long count = reviewRepository.countByBook(book);
-		BigDecimal average = reviewRepository.findAverageRatingByBook(book.getId());
-
-		List<Review> reviews = reviewRepository.findReviewsByBook(book, pageable).getContent();
-		List<ReviewResponseDto> reviewList = reviewMapper.toResponseDto(reviews);
-
-		PaginationDto paginationDto = reviewMapper.toPaginationDto(count, page, size);
-		ReviewSummaryDto summaryDto = reviewMapper.toSummaryDto(count, average);
-
-		ReviewListResponseDto resultResponseDto = reviewMapper.toReviewListResponseDto(reviewList, paginationDto,
-			summaryDto);
-		return BaseResponse.ok("조회가 성공적으로 수행되었습니다.", resultResponseDto, HttpStatus.OK);
 	}
 
 	//파일 받아오기 예시코드
